MAP-Data-Ingestion-Service/requirements.txt
Flask==2.2.2
Flask_RESTful==0.3.9
Flask_Bcrypt==0.7.1
psycopg2==2.9.3
psycopg2-binary==2.9.5
Werkzeug==2.2.2
PyJWT==2.4.0
pytest-postgresql==4.1.1
psycopg==3.1.4
Flask-APScheduler==1.12.4
APScheduler==3.10.0
requests
python-decouple
python-dotenv
nanoid